Add HTML template for emails (missing nice read configuration and UnitTest?)
