+++ IntraWeb/src/IntraWeb/Controllers/Api/v1/BaseController.cs
@@ -1,0 +1,17 @@
+﻿using Microsoft.AspNet.Authorization;
+using Microsoft.AspNet.Mvc;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
+
+namespace IntraWeb.Controllers.Api.v1
+{
+    /// <summary>
+    /// Use this base class for all API controllers.
+    /// </summary>
+    [Authorize]
+    public class BaseController : Controller
+    {
+    }
+}
