--- IntraWeb/src/IntraWeb/IEmailService.cs
@@ -1,6 +1,0 @@
-﻿namespace IntraWeb.Controllers.Api
-{
-    internal interface IEmailService
-    {
-    }
-}
