+++ IntraWeb/src/IntraWeb/Startup.cs
@@ -14,6 +14,9 @@ using IntraWeb.Services;
 using IntraWeb.ViewModels.Administration;
 using Microsoft.AspNet.Authentication.Cookies;
 using System.Net;
+using Microsoft.AspNet.Mvc;
+using Microsoft.AspNet.Authorization;
+using IntraWeb.Filters;
 
 namespace IntraWeb
 {
@@ -76,10 +79,10 @@ namespace IntraWeb
             .AddEntityFrameworkStores<ApplicationDbContext>()
             .AddDefaultTokenProviders();
 
-            services.AddMvc();
+            services.AddMvc(); // ToDo: Replace with Web API when it will be done in ASP.NET Core 1.0
 
             // Add application services
-            services.AddTransient<IEmailSender, EmailService>();
+            services.AddTransient<IEmailService, EmailService>();
 
             //services.AddInstance<IRoomRepository>(new Models.Dummies.RoomDummyRepository()); //Testovacia implementacia
             services.AddScoped<IRoomRepository, RoomsRepository>();
@@ -124,7 +127,7 @@ namespace IntraWeb
             AdministrationModelMapping.ConfigureRoomMapping();
 
             // To configure external authentication please see http://go.microsoft.com/fwlink/?LinkID=532715
-            app.UseMvc();            
+            app.UseMvc();
         }
 
         // Entry point for the application.
